Move stable horses screen to stable ui package
